requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==3.1
django-mysql==3.8.1
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jwt==1.0.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.1.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
